requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==25.3.0
bcrypt==4.2.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
cryptography==43.0.1
databases==0.9.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.0
frozenlist==1.6.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
gspread==6.2.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.4.4
nest-asyncio==1.6.0
oauthlib==3.2.2
passlib==1.7.4
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.3
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.12
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==58.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.38.6
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.31.0
yarl==1.20.0
